requirements.txt
geopy
pandas
xlrd
psycopg2-binary